feat: ✨ update algorithm logic to include regex in parentheses (#159)
## Description

This PR updates the logic in `get_algorithm_logic()` to include regex in
parentheses.
We'll need for e.g., podiatrist services: `(speciale =~ '^54') AND
(barnmak != 0)`.

To test the updated regex, I added an argument to `get_algorithm_logic`
with a default of `algorithm`. This enables me to add a
`test_algorithm_logic` object instead of the algorithm we use, so the
tests won't start failing if/when the algorithm is updated.

I don't think this is good practice (adding an arg mainly just for test
purposes), but I wanted to test the logic formally now that it's a bit
more complex. Let me know if there's a better way to do this!

---------

Co-authored-by: Luke W. Johnston <lwjohnst@gmail.com>
Co-authored-by: Anders Aasted Isaksen <67263135+Aastedet@users.noreply.github.com>
Co-authored-by: Luke W. Johnston <lwjohnst86@users.noreply.github.com>