fix: 🐛 change honuge format to yyww (#165)
This PR changes the format of the variable `honuge` from `wwyy` to
`yyww`, since that's the correct format, according to
[DST](https://www.dst.dk/da/TilSalg/Forskningsservice/Dokumentation/hoejkvalitetsvariable/sygesikring---ydelser/honuge).
I have changed this in the data simulation and variable description.

I have also added a zero padding of the week number, since that's how
it's described by DST (in the link provided above:
> HONUGE er angivet som ÅÅUU, altså 2-cifret årstal efterfulgt af
ugenummer 01 til 52 (evt. 53).

NOTE: this does not update the simulated data. I'll add that to the
targets pipeline in a separate PR and update it there.