docs: :memo: separate algorithm vignette, plus pseudocode for internals (#163)
Building on the wonderful work that @Aastedet started in #133, I started
splitting out the algorithm details from the internal function
implementation details. I also converted a lot of his text into
pseudocode so we can just cut and paste as we develop them.

I'm still ongoing with this, but it's looking amazing so far!!

---------

Co-authored-by: Anders Aasted Isaksen <ANDAAS@onerm.dk>
Co-authored-by: Anders Aasted Isaksen <67263135+Aastedet@users.noreply.github.com>
Co-authored-by: Signe Kirk Brødbæk <40836345+signekb@users.noreply.github.com>