refactor: :recycle: move algorithm into R file (#204)
## Description

It was really annoying to have to make a change to the algorithm,
regenerate the rda by re-running the targets pipeline, then re-running
the code to get the correct algorithm things used. Combined with the
fact that rda is a binary file so if it is touched, it triggers a Git
tracking change. So I moved the algorithm over into R code instead.

Closes #200

## Checklist

- [x] Ran `just run-all`