feat: :sparkles: skeleton of the core diabetes classification (#142)
Closes #141.

This will form the basis of slowly building up this function. So as we
add more inclusion/exclusion functions, we uncomment from here. And at
the same time, the general logic of what is needed and what goes into
what, as well as the argument names of the functions, can be seen from
this function.

@Aastedet I know you had some things you mentioned that needed updating
from the diagram, so maybe this function can help clarify that.

---------

Co-authored-by: Signe Kirk Brødbæk <40836345+signekb@users.noreply.github.com>