refactor: :recycle: moved registers into its own R code (#221)
## Description

I moved the register metadata over from the `data-raw` into it's own R
file. This required some changes to other areas. I also moved some of
the functions that were only used in the `data-sources.Rmd` vignette
into that file, since it is generally bad practice to use `:::`, which
is what I'd done before just to get started. This also resulted in CRAN
checks giving a warning about the functions used there but not in the R
files, so I moved the vignettes into `articles/` because CRAN doesn't
check those files (nor do they get uploaded to CRAN).

Related to #215, this PR prepares it so we can check the variable data
types.

## Checklist

- [x] Ran `just run-all`
- [x] If docs were added, Markdown is formatted

---------

Co-authored-by: Signe Kirk Brødbæk <40836345+signekb@users.noreply.github.com>
Co-authored-by: Anders Aasted Isaksen <67263135+Aastedet@users.noreply.github.com>