feat: :sparkles: uncomment first exclusion step (#243)
## Description

After I uncommented the first exclusion step, there was an error because
the data for date is a string in the form YYYYMMDD. While we said to
assume it is a date, if we also assume the variable is in the original
format, it won't be a date. So this addition ensures the values are
dates.

Closes #218 by not doing what is described in the issue.

## Checklist

- [x] Ran `just run-all`