build: :building_construction: add sections to Reference (#286)
## Description

Thought it might be nice to have some sections in the reference docs so
we have a clearer distinction between what's actually directly related
to the OSDC algorithm and what's related to documentation and examples.

Not entirely sure about the wording of the title and descriptions
though. So, please chime in if you have some suggestions on that 🙌

It looks like this: 


![image](https://github.com/user-attachments/assets/224e61bb-20f8-458b-8d75-181e9bec0cde)

## Checklist

- [X] Ran `just run-all`

---------

Co-authored-by: Luke W. Johnston <lwjohnst86@users.noreply.github.com>