docs: :memo: update tables in algorithm vignette (#292)
## Description

In this PR, I transform everything in the algorithm to a df, including
the variable name of the logic.

I've kept one table per register, but removed the register column to
give more space for the new `name` column and the comments.

I've added a table at the bottom for all logic with `is.na(register)` to
include logic that goes across registers.

Closes #285 

## Checklist

- [X] Ran `just run-all`

---------

Co-authored-by: Luke W. Johnston <lwjohnst86@users.noreply.github.com>