feat: :sparkles: `join_inclusions()` added as step (#246)
## Description

I didn't add example code nor tests because it's starting to get to the
point that there's a lot of coupling going on.

Instead we should now start expanding the tests for
`classify_diabetes()` rather than the other functions, as well as adding
a guide that can run the code as examples (rather than have them in the
example roxygen section).

Closes #244

## Checklist

- [x] Ran `just run-all`