feat: :sparkles: `check_data_types()` (#304)
## Description

This PR adds a `check_data_types()` function with tests (intentionally
not using `verify_` --> I will change the name of
`verify_required_variables()` in a separate PR). This function uses the
data types described in `registers()` and is used within
`classify_diabetes()` to check that the variables have the expected data
types in the initial checks before going into the actual filtering.

In this PR, I've also updated the expectations for `foed_dato` so it can
be both a `Date` or `character`, just like the rest of the date columns.
Lastly, I've updated the simulation definition for `koen` bc it was a
`numeric` otherwise. Alternatively, we could change the expected data
type to be `numeric`.

Thoughts?

Closes #215 
Closes #220 (I think?)

Might close #263 if we agree that expecting either a `Date` or
`character` is adequate for date columns?
If we want to transform columns, are we satisfied with just doing that
in other functions, as `c_spec` is currently being transformed to an
`integer` in `prepare_lpr2()`?

---------

Co-authored-by: Luke W. Johnston <lwjohnst86@users.noreply.github.com>