refactor: ♻️ add separate `keep_two_earliest_events()` function (#369)
## Description

Move the functionality into a separate function, add it in
`classify_diabetes()` and remove it from `keep_podiatrist_services()`
and `add_t1d_diagnoses_cols()`.

This PR also updates the function flow vignette based on this change.

I think it's the cleanest to keep it out of `join_inclusions()` so that
function keeps being about joining only, but let me know if you
disagree.

Closes #347 

## Checklist

- [X] Ran `just run-all`
- [X] If docs were added, Markdown is formatted