docs: :memo: add diagram for general classification description (#374)
## Description

Made in Mermaid, naturally. Mirroring the textual description and
[Anders' earlier
plot](https://raw.githubusercontent.com/Aastedet/dissertation/main/figures/osdc_flow/inclusion_flow.svg)
for this.

This is very much a case of "no one asked you to do this, Signe", but I
started this while waiting for Anders today, and then I thought that I
might as well finish it up and make a PR for it.

Bc I've also been made the merger of everything today, I'll just merge
this without review. @Aastedet, you're ofc welcome to change it when you
review the documentation.

## Checklist

- [X] Ran `just run-all` 
- [X] If docs were added, Markdown is formatted