fix: :bug: convert NAs to FALSE and drop any if TRUE for pregnancy (#398)
## Description

This simplifies the issue we discussed in #379. At the point of
calculating pregnancy interval, I converted all NAs to FALSE, since we
only want those with TRUE to be dropped. Also converted to `any()` as in
the discussion of #379.

I also simplified it to be `is_within_`, since a double (maybe even
triple?) negative can be difficult to understand.

## Checklist

- [x] Ran `just run-all`