refactor: ♻️ add logical inclusion variables to classification logic (#404)
## Description

This PR adds the logical inclusion variables to the `has_t1d` logic,
including that there should be 2 or more of these that are `TRUE`. This
implementation (with >=2) only makes sense if we go with the changes in
#379.

To be more explicit, I've also added separate `classify_t2d()` function
with `has_t2d` logic that includes the logical inclusion variables as
well (and that there should be 2 or more of them that are `TRUE`).

Let me know what you think about this implementation and if this is
aligned with what was intended for #401.

As @Aastedet noted on Discord in regards to the current state of `main`,
this PR also only survives the tests bc the edge case test is skipped.

Closes #401 

## Checklist

- [X] Ran `just run-all`

---------

Co-authored-by: Luke W. Johnston <lwjohnst86@users.noreply.github.com>