build: :hammer: add Jarl linter to justfile and CI (#422)
## Description

The original lintr R package was nice but it was too slow to use nicely
in a dev workflow. This adds the new Jarl linter, written in Rust, that
is super fast. I ran it and there are no issues aside from the stuff in
`dev/` (which will be removed in #421).

Closes #418

## Checklist

- [x] Ran `just run-all`