docs: :truck: reduce build size by converting Mermaid to SVG (#451)
## Description

CRAN requires that HTML files be self-contained, meaning any styling or
theming needs to be included inside of the HTML file itself. When Quarto
makes the Mermaid images, it puts it in as JavaScript with all of its
dependencies included... and its a lot. So the file sizes were like 2 Mb
each, which is why we were getting the NOTE about package size. This
seems to solve that NOTE.

Closes #449

## Checklist

- [x] Ran `just run-all`
- [x] If docs were added, Markdown is formatted